First usrp acquisition in aff3ct module.

+++ src/common/Factory_DVBS2O/Factory_DVBS2O.hpp
@@ -69,7 +69,7 @@ struct Factory_DVBS2O {
 	template <typename B = int>
 	static module::Monitor_BFER<B>* build_monitor(const Params_DVBS2O& params);
 
-	template <typename R = float>
+	template <typename R = double>
 	static module::Radio<R>* build_radio (const Params_DVBS2O& params);	
 };
